requirements.txt
llama-index-embeddings-huggingface==0.5.2
llama-index-vector-stores-qdrant==0.6.0
llama-index-core==0.12.28
llama-index-llms-groq==0.3.1
qdrant-client==1.13.3



